requirements.txt
fastapi
uvicorn
numpy
scikit-learn
sentence-transformers
beautifulsoup4
torch
faiss-cpu